requirements.txt
flask
pytesseract
pillow
pandas
requests
beautifulsoup4
